broker/cloud_run/bq_rest_api/requirements.txt
Flask
flask-marshmallow
Flask-RESTful
pyarrow
python-dotenv
SQLAlchemy
sqlalchemy-bigquery
